Rename bcsd_downscale_archive.py downscale_archive.py
